refactor: rename encrypted package to datastore
