+++ data/src/main/java/com/madrid/data/dataSource/datastore/AuthenticationDatastore.kt
@@ -1,4 +1,4 @@
-package com.madrid.data.dataSource.encrypted
+package com.madrid.data.dataSource.datastore
 
 import android.content.Context
 import androidx.datastore.core.DataStore
